Move exceptions and warnings into alns/tools
