+++ alns/tools/warnings.py
@@ -1,0 +1,6 @@
+class OverwriteWarning(UserWarning):
+    """
+    Raised when the algorithm detects that an operator has the same name
+    as an already existing operator of that type.
+    """
+    pass
